pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "pybind11>=2.10"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-project"
version = "0.1.0"
description = "Local RAG system with Python + C++ chunking."
requires-python = ">=3.11"
dependencies = [
  "typer",
  "fastembed",
  "lancedb",
  "pydantic",
  "pyyaml",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["rag"]
